seshat: Move the index into a separate folder.
