+++ src/index/encrypted_stream.rs
